requirements.txt
git+https://github.com/Azd325/gingerit.git
streamlit==1.32.0
python-docx==1.1.0
transformers==4.41.0
torch==2.3.0
regex==2024.4.28
sentencepiece==0.2.0
protobuf==3.20.3
